Add kernel boot parameter block
The kernel boot parameter block allows the bootloader to pass
binary data to the kernel during boot. The block must be placed
at a formerly agreed-upon location (currently 0x7000). It must
begin with a specified signature word (currently 0xc33c).

The block is placed within the bootloader's uninitialised
section, and must be created entirely at run-time.

Currently the kernel is passed the co-ordinates of the text-mode
cursor, so that text-output can continue from the same on-screen
location - providing continuity·